feat: show loading screen while save data is being fetched
